codigo + organizado

+++ components/common/common_perif.c
@@ -2,7 +2,7 @@
 
 #include <stdio.h>
 
-perif_update_t perif_update_create(perif_event_t type, perif_t target, perif_resp_t resp_type, uint8_t value) {
+perif_update_t perif_update_create(perif_event_t type, perif_t target, perif_resp_t resp_type, int64_t value) {
 	perif_update_t perif_update = (perif_update_t) malloc(sizeof(s_perif_update_t));
 
 	perif_update->type = type;
