+++ components/common/include/common_perif.h
@@ -35,6 +35,6 @@ typedef struct s_perif_update_t {
 	int64_t value;
 } s_perif_update_t;
 
-perif_update_t perif_update_create(perif_event_t type, perif_t target, perif_resp_t resp_type, uint8_t value);
+perif_update_t perif_update_create(perif_event_t type, perif_t target, perif_resp_t resp_type, int64_t value);
 
 #endif
